refactor(core): extract proxy agent creation
